requirements/requirements.txt
# SHA1:d1ef2273e21b520b9bbebd1b8b69d00cc2d42eb3
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
amqp==5.2.0
    # via kombu
billiard==4.2.0
    # via celery
celery[redis]==5.3.6
    # via
    #   -r requirements/requirements.in
    #   celery-redbeat
celery-redbeat==2.2.0
    # via -r requirements/requirements.in
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
kombu==5.3.5
    # via celery
prompt-toolkit==3.0.43
    # via click-repl
python-dateutil==2.8.2
    # via
    #   celery
    #   celery-redbeat
redis==5.0.1
    # via
    #   celery
    #   celery-redbeat
six==1.16.0
    # via python-dateutil
tenacity==8.2.3
    # via celery-redbeat
tzdata==2023.4
    # via celery
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
